Ok on wasm

+++ src/indexer/stamper.rs
@@ -1,15 +1,15 @@
-use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
 
 #[derive(Clone, Default)]
-pub struct Stamper(Arc<AtomicU64>);
+pub struct Stamper(Arc<AtomicUsize>);
 
 impl Stamper {
     pub fn new(first_opstamp: u64) -> Stamper {
-        Stamper(Arc::new(AtomicU64::new(first_opstamp)))
+        Stamper(Arc::new(AtomicUsize::new(first_opstamp as usize)))
     }
 
     pub fn stamp(&self) -> u64 {
-        self.0.fetch_add(1u64, Ordering::SeqCst)
+        self.0.fetch_add(1, Ordering::SeqCst) as u64
     }
 }
